LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/timeline-scope-expected.txt


PASS Descendant can attach to deferred timeline
PASS Deferred timeline with no attachments
PASS Inner timeline does not interfere with outer timeline
PASS Deferred timeline with two attachments
PASS Dynamically re-attaching
FAIL Dynamically detaching assert_equals: expected "100px" but got "0px"
FAIL Removing/inserting element with attaching timeline assert_equals: expected "0px" but got "100px"
PASS Ancestor attached element becoming display:none/block
FAIL A deferred timeline appearing dynamically in the ancestor chain assert_equals: expected "100px" but got "0px"
PASS Animations prefer non-deferred timelines
